backend/requirements.txt
fastapi==0.112.1
uvicorn[standard]==0.30.5
python-multipart==0.0.9
pandas==2.2.2
pydantic==2.11.7
pdfplumber==0.11.4
openai==1.100.2
httpx==0.28.1
aiofiles==23.2.1
ulid-py==1.1
